NMS-1345: Move HealthCheck test from FlowStackJms to new HealthCheckIT
